requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Faker==30.3.0
Flask==3.0.3
fonttools==4.54.1
inexactsearch==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Levenshtein==0.26.0
MarkupSafe==2.1.5
matplotlib==3.9.2
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.2
RapidFuzz==3.10.0
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
silpa_common==0.3
six==1.16.0
soundex==1.1.3
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
wordcloud==1.9.3
